+++ CHANGES.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Rename Test.During to Test.DuringDefault
+- Rename Test.Report to Test.AfterDefault
+
 ## 0.4.0 - 2023-12-20
 
 ### Added